% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotdrg.R
\name{boxplotdrg}
\alias{boxplotdrg}
\title{boxplot of payments by DRG code}
\usage{
boxplotdrg(var_y)
}
\arguments{
\item{var_y}{}
}
\value{
plot
}
\description{
boxplot of payments by DRG code
}
\examples{
boxplotdrg("Average.Medicare.Payments")

}
