% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calculate.R
\name{stat_calculate}
\alias{stat_calculate}
\title{stat_calculate}
\usage{
stat_calculate(stat = c("mean", "median", "standard_deviation"))
}
\arguments{
\item{stat}{}
}
\value{
df
}
\description{
stat_calculate
}
\examples{
stat_calculate("mean")
}
